pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roughvol"
version = "0.1.0"
description = "Research-grade rough volatility pricing framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "scipy",
  "pandas",
  "matplotlib",
]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
]
grpc = [
  "grpcio>=1.60",
  "grpcio-tools>=1.60",
]

[project.scripts]
rough-pricing-server = "roughvol.service.server:main"
